Separate modal content: make one Modal component and two different contents (with form and with preview content), create new component BookPreview, rename content dives to root dives in modal content components
