+++ components/AddBookGroup/AddBookGroup.jsx
@@ -5,7 +5,8 @@ import { useState } from 'react';
 import s from './addBookGroup.module.scss';
 
 import Button from '../Button';
-import CreateModal from '../CreateModal';
+// import CreateModal from '../CreateModal';
+import Modal from '../Modal';
 
 export default function addBookGroup() {
   const [openModal, setOpenModal] = useState(false);
@@ -19,7 +20,7 @@ export default function addBookGroup() {
       <p className={`${s.arrowHeader} ${caveat.className}`}>Тыкай сюда чтобы добавить книгу</p>
       <img className={s.arrow} src="/arrow.svg" alt="arrow" />
       <Button img="/buttonAdd.svg" alt="add book button" onClick={showModal} className={s.addBtn} />
-      {openModal && <CreateModal onClick={setOpenModal} />}
+      {openModal && <Modal modalType="show create form" onClick={setOpenModal} />}
     </div>
   );
 }
