+++ components/Book/Book.jsx
@@ -2,7 +2,6 @@
 
 import { useState } from 'react';
 import { number, string } from 'prop-types';
-import cx from 'classnames';
 import s from './book.module.scss';
 import Modal from '../Modal';
 import Button from '../Button';
@@ -17,8 +16,8 @@ export default function Book({
   };
 
   return (
-    <div className={cx(s.root, { [s.modal]: openModal })}>
-      {openModal && <Modal title={title} author={author} year={year} onClick={setOpenModal} />}
+    <div className={s.root}>
+      {openModal && <Modal modalType="show book" title={title} author={author} year={year} onClick={setOpenModal} />}
       <div className={s.header}>
         <h2>{title}</h2>
       </div>
