+++ components/Book/book.module.scss
@@ -27,6 +27,7 @@
 
   h2 {
     font-weight: bold;
+    word-break: break-all
   }
 }
